Move NuGetMetadataFactory to MetadataReferences
